Updated blog: ocean physics
