--- pages/blog/kelvin_wave_infographic/index.module.css
@@ -1,28 +1,0 @@
-.canvas {
-  min-height: 80vh;
-  align-items: center;
-}
-
-.textBg {
-  width: 100%;
-  max-width: 700px;
-  margin: auto;
-  background-color: #eee8d5;
-  margin-bottom: 1rem;
-}
-
-.textCanvas {
-  white-space: pre-line;
-  width: 90%;
-  max-width: 45em;
-  margin: auto;
-  padding-top: 0.1rem;
-  padding-bottom: 0.1rem;
-  background-color: #eee8d5;
-  font-family: Arvo;
-}
-
-
-.backIcon:hover {
-  cursor: pointer;
-}
